refactor(rimap-imap): split connection.rs into handshake/login/dispatch
Splits the 1863-LOC connection.rs into a connection/ directory
module:
  - mod.rs (475): Connection struct, ConnectionConfig, ImapSession,
    Debug impl, enrich_tls_handshake_error, Connection lifecycle
    methods (new, host, username, session, invalidate, connect_inner,
    has_*_capability), plus tests for enrich_tls_handshake_error and
    error_code_for.
  - handshake.rs (681): connect_with_bundle method, TCP/TLS/STARTTLS
    free functions (capability_advertised, drain_for_logindisabled,
    starttls_negotiate, drain_for_starttls, tls_handshake,
    starttls_upgrade, map_tls_handshake_error), plus tests for the
    full STARTTLS negotiation path + map_tls_handshake_error +
    Connection's Debug impl.
  - login.rs (339): imap_login + emit_auth methods plus the cancel-
    safety test for emit_auth.
  - dispatch.rs (433): with_session helper + 15 per-command wrappers
    (list_folders, select, search, fetch, fetch_body, store_flags,
    move_messages, append_message, delete_message, expunge,
    create_folder, rename_folder, delete_folder).

Bumped `Connection.inner` and `ConnectionInner` (and its fields) to
`pub(super)` so sibling files can access them. `pub(crate) use
handshake::{starttls_upgrade, tls_handshake};` preserves
preflight.rs's import path. No public-API change.

No behavior change. All 1416 workspace tests still pass.

Refs #300 (item 1)
